+++ packages/bruno-cli/src/utils/common.js
@@ -1,4 +1,5 @@
 const iconv = require('iconv-lite');
+const { execSync } = require('child_process');
 
 const lpad = (str, width) => {
   let paddedStr = str;
@@ -42,8 +43,38 @@ const parseDataFromResponse = (response, disableParsingResponseJson = false) => 
   return { data, dataBuffer };
 };
 
+const splitCsv = (value) =>
+  value
+    ? String(value)
+        .split(',')
+        .map((entry) => entry.trim())
+        .filter(Boolean)
+    : [];
+
+const findConflict = (include, exclude) => {
+  const excluded = new Set(exclude);
+  return include.find((name) => excluded.has(name));
+};
+
+const getGitRemoteUrl = (collectionPath) => {
+  try {
+    const url = execSync('git remote get-url origin', {
+      cwd: collectionPath,
+      stdio: ['ignore', 'pipe', 'ignore']
+    })
+      .toString()
+      .trim();
+    return url || undefined;
+  } catch (error) {
+    return undefined;
+  }
+};
+
 module.exports = {
   lpad,
   rpad,
-  parseDataFromResponse
+  parseDataFromResponse,
+  splitCsv,
+  findConflict,
+  getGitRemoteUrl
 };
